Fix Spring Boot 2 Java 8 compatibility
