+++ smoke-tests-otel-starter/spring-boot-3/build.gradle.kts
@@ -21,6 +21,7 @@ dependencies {
   implementation(platform(org.springframework.boot.gradle.plugin.SpringBootPlugin.BOM_COORDINATES))
 
   implementation(project(":smoke-tests-otel-starter:spring-boot-common"))
+  implementation(project(":instrumentation:runtime-telemetry:runtime-telemetry-java17:library"))
   testImplementation("org.testcontainers:junit-jupiter")
   testImplementation("org.testcontainers:kafka")
   testImplementation("org.testcontainers:mongodb")
